Fix hidden z-values in schedule plot
In general, instead of the plot just using the job start and end, one
can also explicitly plot the z values to see if the constraints are
valid (i.e. all z values are 1 only within the start and end time).
To plot this, just the edges of the rectangle showing the start and end
times should be rendered.
However, in this case, since it still used a solid color, it hid the
bars from the z values behind it.
So if a z value inside of the start to end time would be 0, there
should have been a white gap visible where in reality, another solid
bar was just drawn on top.